Moved get_characters to a new file.
Allowed conversation to have only one person
