+++ utils.py
@@ -1,0 +1,35 @@
+from constants import Character
+import random
+from collections import Counter
+
+def get_characters(common: Counter):
+    users_to_characters = {}
+    most_common =  [t[0] for t in common.most_common()]
+    all_rnd_characters = [
+        Character.GODOT,
+        Character.FRANZISKA,
+        Character.JUDGE,
+        Character.LARRY,
+        Character.MAYA,
+        Character.KARMA,
+        Character.PAYNE,
+        Character.MAGGEY,
+        Character.PEARL,
+        Character.LOTTA,
+        Character.GUMSHOE,
+        Character.GROSSBERG,
+    ]
+    rnd_characters = []
+    if len(most_common) > 0:
+        users_to_characters[most_common[0]] = Character.PHOENIX
+        if len(most_common) > 1:
+            users_to_characters[most_common[1]] = Character.EDGEWORTH
+            for character in most_common[2:]:
+                if len(rnd_characters) == 0:
+                    rnd_characters = all_rnd_characters.copy()
+                rnd_character = random.choice(
+                    rnd_characters
+                )
+                rnd_characters.remove(rnd_character)
+                users_to_characters[character] = rnd_character
+    return users_to_characters
